Adding new section about karate tests to rancher scrach env doc
Running the karate integration tests in the rancher scratch environment requires some
special considerations. The supertenant needs to be secured, and a special user needs to be
created. This documents how to do each of those things as well as adds some advice for
'when things go wrong' based on experience running these tests in the rancher scratch
envs.